ADD WORDPRESS SITE INITIALISATION BASH SCRIPT
This script used the WP CLI command to configure the WordPress
installation with some of the specific requirements of this website. At
the moment it reflects the fact that I've been running with via the
newopera-docker repository. It needs to be generalised so that the same
script can be used also for Ansible deployments of this website. This
approach may come to replace some aspects of initialising WordPress
sites that are currently encapsulated within the wordpress role of the
libraries-ansible repository.